adding confirm message to delete button

+++ src/services/sculptures-services.js
@@ -8,10 +8,11 @@ export const getSculptures = async () => {
 
 // DELETE
 export const deleteSculptures = async (id) => {
-    const response = await fetch(`http://localhost:3000/sculptures/${id}`, {
-        method: 'DELETE'
-    });
-    return response;
+    if (confirm("¿Estás seguro que quieres eliminar?") === true) {
+        const response = await fetch(`http://localhost:3000/sculptures/${id}`, {
+            method: 'DELETE'
+        });
+    }
 };
 
 // POST
